requirements.txt
pytesseract
pdf2image
Pillow
transformers
torch
python-dotenv
google-generativeai
streamlit
jiwer
scikit-learn
rouge-score
bert-score
textstat
requests
fpdf
